Add an exclude flag
The exclude flag (`-x` or `--exclude`) will add a file glob pattern to a
set of excludes. When excludes are specified, paths that are watched
will only trigger live reloads if one of the modified files does not
match the exclude patterns.

A unit tests has been added for file excluding to ensure that it behaves
correctly.

This patch fixes #8.